Fix: Separate settings menu tabs using events
